Move dispatcher types to runtime
