--- exonum/src/proto/schema/exonum/dispatcher.proto
@@ -1,43 +1,0 @@
-// Copyright 2019 The Exonum Team
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//   http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-// Internal dispatcher data types.
-
-syntax = "proto3";
-
-package exonum.dispatcher;
-
-option java_package = "com.exonum.core.messages";
-
-import "runtime.proto";
-
-enum ArtifactStatus {
-    ARTIFACT_PENDING = 0;
-    ARTIFACT_ACTIVE = 1;
-}
-
-enum InstanceStatus {
-    SERVICE_PENDING = 0;
-    SERVICE_ACTIVE = 1;
-}
-
-message ArtifactState {
-    exonum.runtime.ArtifactSpec spec = 1;
-    ArtifactStatus status = 2; 
-}
-
-message InstanceState {
-    exonum.runtime.InstanceSpec spec = 1;
-    InstanceStatus status = 2;
-}
