+++ exonum/src/runtime/types.rs
@@ -343,6 +343,110 @@ impl<'a> From<&'a str> for InstanceQuery<'a> {
     }
 }
 
+/// Status of an artifact deployment.
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
+pub enum ArtifactStatus {
+    /// The artifact is pending deployment.
+    Pending = 0,
+    /// The artifact has been successfully deployed.
+    Active = 1,
+}
+
+impl Display for ArtifactStatus {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            Self::Active => f.write_str("active"),
+            Self::Pending => f.write_str("pending"),
+        }
+    }
+}
+
+impl ProtobufConvert for ArtifactStatus {
+    type ProtoStruct = schema::runtime::ArtifactStatus;
+
+    fn to_pb(&self) -> Self::ProtoStruct {
+        match self {
+            Self::Active => Self::ProtoStruct::ARTIFACT_ACTIVE,
+            Self::Pending => Self::ProtoStruct::ARTIFACT_PENDING,
+        }
+    }
+
+    fn from_pb(pb: Self::ProtoStruct) -> Result<Self, failure::Error> {
+        Ok(match pb {
+            Self::ProtoStruct::ARTIFACT_ACTIVE => Self::Active,
+            Self::ProtoStruct::ARTIFACT_PENDING => Self::Pending,
+        })
+    }
+}
+
+/// Status of a service instance.
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
+pub enum InstanceStatus {
+    /// The service instance is pending deployment.
+    Pending = 0,
+    /// The service instance has been successfully deployed.
+    Active = 1,
+}
+
+impl Display for InstanceStatus {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            Self::Active => f.write_str("active"),
+            Self::Pending => f.write_str("pending"),
+        }
+    }
+}
+
+impl ProtobufConvert for InstanceStatus {
+    type ProtoStruct = schema::runtime::InstanceStatus;
+
+    fn to_pb(&self) -> Self::ProtoStruct {
+        match self {
+            Self::Active => Self::ProtoStruct::SERVICE_ACTIVE,
+            Self::Pending => Self::ProtoStruct::SERVICE_PENDING,
+        }
+    }
+
+    fn from_pb(pb: Self::ProtoStruct) -> Result<Self, failure::Error> {
+        Ok(match pb {
+            Self::ProtoStruct::SERVICE_ACTIVE => Self::Active,
+            Self::ProtoStruct::SERVICE_PENDING => Self::Pending,
+        })
+    }
+}
+
+/// Current state of artifact in dispatcher.
+#[derive(Debug, Clone, PartialEq, Eq, Hash, ProtobufConvert, BinaryValue, ObjectHash)]
+#[protobuf_convert(source = "schema::runtime::ArtifactState")]
+pub struct ArtifactState {
+    /// Artifact specification.
+    pub spec: ArtifactSpec,
+    /// Artifact deployment status.
+    pub status: ArtifactStatus,
+}
+
+impl From<ArtifactState> for (ArtifactStatus, ArtifactSpec) {
+    fn from(v: ArtifactState) -> Self {
+        (v.status, v.spec)
+    }
+}
+
+/// Current state of service instance in dispatcher.
+#[derive(Debug, Clone, PartialEq, Eq, Hash, ProtobufConvert, BinaryValue, ObjectHash)]
+#[protobuf_convert(source = "schema::runtime::InstanceState")]
+pub struct InstanceState {
+    /// Service instance specification.
+    pub spec: InstanceSpec,
+    /// Service instance activity status.
+    pub status: InstanceStatus,
+}
+
+impl From<InstanceState> for (InstanceStatus, InstanceSpec) {
+    fn from(v: InstanceState) -> Self {
+        (v.status, v.spec)
+    }
+}
+
 #[test]
 fn parse_artifact_id_correct() {
     "0:my-service:1.0.0".parse::<ArtifactId>().unwrap();
